refactor(web): extract shared source and result viewer panels
